requirements.txt
groq==0.5.0
rich==13.7.1
python-dotenv==1.0.1
